Simplify code and support urlencoded string
The problem arose where we have people adding variables into the URL. The problem is that generally these are URL encoded before being sent to the server, which is correct, really. The problem is that our current regular expression won't detect this. When trying to update this I found the code quite difficult to follow and set about cleaning it up.

I find this change to be much cleaner and easier to follow/read. There may be a slight performance cost but it should be negligable and unimportant in this case for realistic usage. The regex should also be easier to follow and less error prone as the negative lookup was not necessary. The controversial part of this PR may be that it now includes the URLEncoded version of the identifier in the expression. There are other ways to solve this but I decided on this one. The cleanup could be viewed separate to this if we didn't want to go ahead with this solution for the URL encoding.